Feat: [feat/Integrate] -viewModel의 waterQuantities의 데이터 타입 변경 -버튼 스타일 수정

+++ Models/ViewModel.swift
@@ -30,7 +30,7 @@ struct DripSessionModel {
     var Dripper: String = ""
     var beanAmount: Int = 0
     var pourTimeSums: [Double] = []
-    var waterQuantities: [Double] = []
+    var waterQuantities: [Int] = []
 }
 
 class ViewModel: ObservableObject {
